requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.2.0
docx2txt>=0.8
python-dotenv>=1.0.0
